Bug-fix: The react app was not able to run in docker
Bug-cause: The react app was not able to run in docker as it requires interactive mode to run.
Solution: The bug was fixes by adding line 'stdin_open: true' in 'docker-compose' file.
This line enables the docker container to run in interactive mode which is necessary for react app to run in development mode.
This is an issue with 'react-scripts@3.4.1' it will also work if we downgrade the 'react-scripts@3.4.1' to lower version.
The full discussion for this solution was found here (stackoverflow.com): https://stackoverflow.com/questions/61837934/content-not-from-webpack-is-served-from-app-public-docker.